PWN-4512 - Fix Token List Diff Utils (#642)
* PWN-4396 - Add new swipe layout for adapter

* PWN-4396 - Clear branch

* PWN-4396 - Resolve all MR comments

* PWN-4396 - Resolve all MR comments

* PWN-4396 - Fix bottom navigation, and add Diffs for token list

* PWN-4396-fix ktlint

* PWN-4396 - Remove alpha from token list

* PWN-4396 - Fix hide state of token list
