+++ app/src/main/res/layout/fragment_main.xml
@@ -22,5 +22,22 @@
         app:labelVisibilityMode="labeled"
         app:layout_constraintBottom_toBottomOf="parent" />
 
+    <org.p2p.uikit.components.UiKitButton
+        android:id="@+id/buttonCenterAction"
+        android:layout_width="60dp"
+        android:layout_height="60dp"
+        android:layout_gravity="center_vertical|center_horizontal"
+        android:layout_marginBottom="8dp"
+        android:background="@drawable/bg_circle"
+        android:scaleType="center"
+        app:backgroundTint="@color/bg_lime"
+        app:icon="@drawable/ic_lightling"
+        app:iconGravity="textTop"
+        app:iconPadding="0dp"
+        app:iconTint="@color/icons_night"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
