+++ app/src/main/res/layout/item_home_token_list_item_shimmer.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingHorizontal="16dp"
+    android:paddingVertical="8dp">
+
+    <ImageView
+        android:id="@+id/iconView"
+        android:layout_width="44dp"
+        android:layout_height="44dp"
+        android:background="@drawable/bg_stroke_rounded"
+        android:backgroundTint="@color/bg_rain"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+
+    <ImageView
+        android:id="@+id/topLeftView"
+        android:layout_width="100dp"
+        android:layout_height="12dp"
+        android:layout_marginHorizontal="16dp"
+        android:background="@drawable/bg_stroke_rounded"
+        android:backgroundTint="@color/bg_rain"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0"
+        android:layout_marginTop="6dp"
+        app:layout_constraintStart_toEndOf="@id/iconView"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintWidth_percent="0.2" />
+
+    <ImageView
+        android:id="@+id/bottomLeftView"
+        android:layout_width="100dp"
+        android:layout_height="12dp"
+        android:layout_marginHorizontal="16dp"
+        android:background="@drawable/bg_stroke_rounded"
+        android:backgroundTint="@color/backgroundAppStroke"
+        app:layout_constraintBottom_toBottomOf="@id/iconView"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginBottom="6dp"
+        app:layout_constraintHorizontal_bias="0"
+        app:layout_constraintStart_toEndOf="@id/iconView"
+        app:layout_constraintWidth_percent="0.3" />
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
+
