+++ ui-kit/src/main/kotlin/org/p2p/uikit/components/UiKitBottomNavigationView.kt
@@ -3,7 +3,7 @@ package org.p2p.uikit.components
 import android.content.Context
 import android.util.AttributeSet
 import android.view.Menu
-import androidx.constraintlayout.widget.ConstraintLayout
+import androidx.coordinatorlayout.widget.CoordinatorLayout
 import org.p2p.uikit.R
 import org.p2p.uikit.databinding.WidgetBottomNavigationViewBinding
 import org.p2p.uikit.utils.inflateViewBinding
@@ -12,7 +12,7 @@ class UiKitBottomNavigationView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0
-) : ConstraintLayout(context, attrs, defStyleAttr) {
+) : CoordinatorLayout(context, attrs, defStyleAttr) {
 
     val binding = inflateViewBinding<WidgetBottomNavigationViewBinding>()
 
@@ -29,10 +29,6 @@ class UiKitBottomNavigationView @JvmOverloads constructor(
         binding.bottomNavigationView.setOnItemSelectedListener { menuItem ->
             ScreenTab.fromTabId(menuItem.itemId)?.let { block.invoke(it) } ?: false
         }
-    }
-
-    fun setOnCenterActionButtonListener(block: () -> Unit) {
-        binding.buttonCenterAction.setOnClickListener { block.invoke() }
     }
 
     fun inflateMenu(menuResId: Int) {
