+++ ui-kit/src/main/res/layout/widget_bottom_navigation_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
+<merge
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -25,20 +25,4 @@
             app:menu="@menu/menu_ui_kit_bottom_navigation" />
     </FrameLayout>
 
-    <org.p2p.uikit.components.UiKitButton
-        android:id="@+id/buttonCenterAction"
-        android:layout_width="60dp"
-        android:layout_height="60dp"
-        android:layout_marginBottom="8dp"
-        android:background="@drawable/bg_circle"
-        android:scaleType="center"
-        app:backgroundTint="@color/bg_lime"
-        app:icon="@drawable/ic_lightling"
-        app:iconGravity="textTop"
-        app:iconPadding="0dp"
-        app:iconTint="@color/icons_night"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+</merge>
